2.3 Кэширование (Вариант А): CachedNetworkImage, photo_cache_manager, flutter_cache_manager, main.cpp fix (Android + Windows)

+++ Windows/windows/runner/main.cpp
@@ -23,16 +23,16 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
     const wchar_t* subKey = L"SOFTWARE\\Classes\\cactusapp";
     if (RegCreateKeyExW(HKEY_CURRENT_USER, subKey, 0, nullptr, 0, KEY_WRITE, nullptr, &hKey, nullptr) == ERROR_SUCCESS) {
       const wchar_t* defaultValue = L"URL:cactusapp Protocol";
-      RegSetValueExW(hKey, nullptr, 0, REG_SZ, (const BYTE*)defaultValue, (wcslen(defaultValue) + 1) * sizeof(wchar_t));
+      RegSetValueExW(hKey, nullptr, 0, REG_SZ, (const BYTE*)defaultValue, static_cast<DWORD>((wcslen(defaultValue) + 1) * sizeof(wchar_t)));
       const wchar_t* urlProtocol = L"";
-      RegSetValueExW(hKey, L"URL Protocol", 0, REG_SZ, (const BYTE*)urlProtocol, (wcslen(urlProtocol) + 1) * sizeof(wchar_t));
+      RegSetValueExW(hKey, L"URL Protocol", 0, REG_SZ, (const BYTE*)urlProtocol, static_cast<DWORD>((wcslen(urlProtocol) + 1) * sizeof(wchar_t)));
 
       HKEY hCmdKey;
       if (RegCreateKeyExW(hKey, L"shell\\open\\command", 0, nullptr, 0, KEY_WRITE, nullptr, &hCmdKey, nullptr) == ERROR_SUCCESS) {
         wchar_t exePath[MAX_PATH];
         GetModuleFileNameW(nullptr, exePath, MAX_PATH);
         std::wstring cmd = std::wstring(exePath) + L" \"%1\"";
-        RegSetValueExW(hCmdKey, nullptr, 0, REG_SZ, (const BYTE*)cmd.c_str(), (cmd.length() + 1) * sizeof(wchar_t));
+        RegSetValueExW(hCmdKey, nullptr, 0, REG_SZ, (const BYTE*)cmd.c_str(), static_cast<DWORD>((cmd.length() + 1) * sizeof(wchar_t)));
         RegCloseKey(hCmdKey);
       }
       RegCloseKey(hKey);
